Validation scripts.
Parallelize the validation by introducing a scheduler-worker structure.

The scheduler determines the configurations that need to be validated. Then, worker can contact the scheduler to get a task to validate.
A worker, after successfully computing a result dictionary, sends the results back to the scheduler.
The scheduler registers the result.