Tackle Idling worker
1 of 10 optimization runs was frozen at the end of the procedure.

Increasing the time between the server pings seems to work. I have also changed the lock, since I assume that there was an interaction between terminating the optimization process and releasing locks.
Also, I improved the shutdown procedure of the container. It may be useful to upgrade also the HPOBench Version!