Parallelize the validation process (#24)
* Add info in validated trajectory

* Add more info in validated trajectory

* Update Logging.

* ADD some more notes and references (#17)

* Validation scripts.

Parallelize the validation by introducing a scheduler-worker structure.

The scheduler determines the configurations that need to be validated. Then, worker can contact the scheduler to get a task to validate.
A worker, after successfully computing a result dictionary, sends the results back to the scheduler.
The scheduler registers the result.

* Validation scripts. Part 2

Integrate the parallelization into the Validation Procedure

* Small fix

Had some troubles with the logging.
Now, we need to call basicConfig only once and can derive in each file the logger from the root logger.

* Clean up

* Small refactoring

* Fix an Issue for installing ConfigSpace

I had to decrease the Configspace version since it was not able to install ConfigSpace with python 3.6.
It did not find the version 0.4.15.

* Tackle Idling worker

1 of 10 optimization runs was frozen at the end of the procedure.

Increasing the time between the server pings seems to work. I have also changed the lock, since I assume that there was an interaction between terminating the optimization process and releasing locks.
Also, I improved the shutdown procedure of the container. It may be useful to upgrade also the HPOBench Version!

* Add installation guide for DEHB

* DEHB Version++

Optimizer can now work with new DEHB version

* Remove the nameserver address file of an old nameserver before starting the new nameserver

* Fix Connection Problems

Note that the signature has changed!
The interface name is now also required when calling the worker.

* Add info in validated trajectory

* Validation scripts.

Parallelize the validation by introducing a scheduler-worker structure.

The scheduler determines the configurations that need to be validated. Then, worker can contact the scheduler to get a task to validate.
A worker, after successfully computing a result dictionary, sends the results back to the scheduler.
The scheduler registers the result.

* Validation scripts. Part 2

Integrate the parallelization into the Validation Procedure

* Small fix

Had some troubles with the logging.
Now, we need to call basicConfig only once and can derive in each file the logger from the root logger.

* Clean up

* Small refactoring

* Fix an Issue for installing ConfigSpace

I had to decrease the Configspace version since it was not able to install ConfigSpace with python 3.6.
It did not find the version 0.4.15.

* Tackle Idling worker

1 of 10 optimization runs was frozen at the end of the procedure.

Increasing the time between the server pings seems to work. I have also changed the lock, since I assume that there was an interaction between terminating the optimization process and releasing locks.
Also, I improved the shutdown procedure of the container. It may be useful to upgrade also the HPOBench Version!

* Add installation guide for DEHB

* DEHB Version++

Optimizer can now work with new DEHB version

* Remove the nameserver address file of an old nameserver before starting the new nameserver

* Fix Connection Problems

Note that the signature has changed!
The interface name is now also required when calling the worker.

* Fix Merge Mistake

* ADD script to collect stats

* FIX typo

Co-authored-by: Katharina Eggensperger <eggenspk@informatik.uni-freiburg.de>