Fix Time Counting in bookkeeper.py
In case the benchmark is a surrogate, we are only interessted in the predicted costs for the objective function. Thus, we subtract the time needed to call the objective function. However, in  the previous versions, we only subtracted the time of the last call and not the summed up time from the calls before.